clean python benches
